Add redux to widget
