+++ packages/moocfi-quizzes/package.json
@@ -25,5 +25,9 @@
     "react": ">=16.0.0",
     "react-dom": ">=16.0.0",
     "styled-components": ">=4"
+  },
+  "dependencies": {
+    "react-redux": "^7.1.0",
+    "redux": "^4.0.1"
   }
 }
