+++ packages/moocfi-quizzes/src/state/store.ts
@@ -1,0 +1,8 @@
+import { createStore } from "redux"
+import rootReducer from "./rootReducer"
+
+const createStoreInstance = () => {
+  return createStore(rootReducer)
+}
+
+export default createStoreInstance
